requirements.txt
fastapi==0.89.1
pillow==8.4.0
pickle5==0.0.11
torch==2.0.1
clip @ git+https://github.com/openai/CLIP.git
